- made fail and success behaviour usable via the TestBehaviourCreator
