--- alica_test_utility/include/alica/test/SuccessBehaviour.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "IMockUpBehaviour.h"
-
-namespace alica::test
-{
-class SuccessBehaviour : public IMockUpBehaviour
-{
-    explicit SuccessBehaviour(uint32_t iterationsBeforeSuccess = 0);
-    void run(void* msg) override;
-private:
-    uint32_t _iterationsBeforeSuccess;
-};
-} // namespace alica::mockups
