--- alica_test_utility/src/SuccessBehaviour.cpp
@@ -1,16 +1,0 @@
-#include "alica/test/SuccessBehaviour.h"
-namespace alica::test
-{
-
-SuccessBehaviour::SuccessBehaviour(uint32_t iterationsBeforeSuccess)
-        : IMockUpBehaviour("SuccesBehaviour")
-        , _iterationsBeforeSuccess(iterationsBeforeSuccess){};
-
-void SuccessBehaviour::run(void* msg)
-{
-    if (iterationsCounter() >= _iterationsBeforeSuccess && !this->isSuccess()) {
-        this->setSuccess();
-    }
-    incIterationsCounter();
-}
-} // namespace alica::mockups
